Move ClusterDetails to pages directory
