+++ assets/js/components/ExecutionResults/TargetInfoBox.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { TARGET_CLUSTER, TARGET_HOST } from '@lib/model';
 
-import { ClusterInfoBox } from '@components/ClusterDetails';
+import { ClusterInfoBox } from '@pages/ClusterDetails';
 import HostInfoBox from '@components/HostDetails/HostInfoBox';
 
 function TargetInfoBox({ targetType, target }) {
