+++ assets/js/pages/ClusterDetails/AscsErsClusterDetails.stories.jsx
