+++ assets/js/pages/ClusterDetails/AttributesDetails.jsx
