+++ assets/js/pages/ClusterDetails/ClusterInfoBox.test.jsx
