+++ assets/js/pages/ClusterDetails/ClusterNodeLink.test.jsx
@@ -3,7 +3,7 @@ import { screen, render, waitFor, act } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import '@testing-library/jest-dom';
 import { renderWithRouter } from '@lib/test-utils';
-import ClusterNodeLink from '@components/ClusterDetails/ClusterNodeLink';
+import ClusterNodeLink from '@pages/ClusterDetails/ClusterNodeLink';
 import { hostFactory } from '@lib/test-utils/factories';
 
 describe('ClusterNodeLink', () => {
