+++ assets/js/pages/ClusterDetails/ExecutionIcon.jsx
