+++ assets/js/pages/ClusterDetails/HanaClusterDetails.jsx
@@ -9,7 +9,7 @@ import BackButton from '@components/BackButton';
 import ListView from '@common/ListView';
 import Table from '@common/Table';
 import Tooltip from '@common/Tooltip';
-import ClusterNodeLink from '@components/ClusterDetails/ClusterNodeLink';
+import ClusterNodeLink from '@pages/ClusterDetails/ClusterNodeLink';
 import CheckResultsOverview from '@components/CheckResultsOverview';
 import ProviderLabel from '@components/ProviderLabel';
 import SapSystemLink from '@components/SapSystemLink';
