+++ assets/js/pages/ClusterDetails/HanaClusterDetails.stories.jsx
