+++ assets/js/pages/ClusterDetails/StoppedResources.jsx
